+++ NEWS.md
@@ -1,3 +1,6 @@
+## Current development version
+
+
 # CBTF 0.6.0 (2026-02-13)
 
 ## New features